Set up unit tests for both server and client
The client came with a mostly working test harness using mocha and chai.
That is now fully working, and the server is set up on the same
toolchain.
