+++ package.json
@@ -16,6 +16,9 @@
   },
   "scripts": {
     "start": "cd packages/client && yarn start",
+    "test": "yarn run test:server && yarn run test:client",
+    "test:client": "cd packages/client && yarn test",
+    "test:server": "cd packages/server && yarn test",
     "format": "prettier '**/*.{ts,tsx,json,graphql}' --write"
   }
 }
