+++ packages/server/package.json
@@ -4,7 +4,8 @@
   "license": "MIT",
   "main": "src/server.ts",
   "scripts": {
-    "start": "ts-node-dev -r tsconfig-paths/register src/server.ts"
+    "start": "ts-node-dev -r tsconfig-paths/register src/server.ts",
+    "test": "NODE_ENV=test ts-mocha \"src/**/*.test.ts\""
   },
   "dependencies": {
     "@graphql-modules/core": "^0.7.17",
@@ -16,8 +17,13 @@
     "winston": "^3.3.3"
   },
   "devDependencies": {
+    "@types/chai": "^4.2.14",
     "@types/express": "^4.17.9",
     "@types/express-winston": "^4.0.0",
+    "@types/mocha": "^8.0.4",
+    "chai": "^4.2.0",
+    "mocha": "^8.2.1",
+    "ts-mocha": "^8.0.0",
     "ts-node-dev": "^1.0.0",
     "tsconfig-paths": "^3.8.0"
   }
